+++ src/main/java/com/hostal/springboot/app/repository/ReservaHabitacionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hostal.springboot.app.model.ReservaHabitacion;
 
-public interface ReservaHabitacionRepository extends CrudRepository<ReservaHabitacion, Long> {
+public interface ReservaHabitacionRepository extends CrudRepository<ReservaHabitacion, Integer> {
 
 }
