+++ src/main/java/com/hostal/springboot/app/repository/ServicioHabitacionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hostal.springboot.app.model.ServicioHabitacion;
 
-public interface ServicioHabitacionRepository extends CrudRepository<ServicioHabitacion, Long> {
+public interface ServicioHabitacionRepository extends CrudRepository<ServicioHabitacion, Integer> {
 
 }
