+++ src/main/java/com/hostal/springboot/app/repository/ServicioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hostal.springboot.app.model.Servicio;
 
-public interface ServicioRepository extends CrudRepository<Servicio, Long> {
+public interface ServicioRepository extends CrudRepository<Servicio, Integer> {
 
 }
