+++ src/main/java/com/hostal/springboot/app/repository/TipoPagoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hostal.springboot.app.model.TipoPago;
 
-public interface TipoPagoRepository extends CrudRepository<TipoPago, Long> {
+public interface TipoPagoRepository extends CrudRepository<TipoPago, Integer> {
 
 }
