+++ src/main/java/com/hostal/springboot/app/services/HuespedService.java
@@ -3,6 +3,6 @@ package com.hostal.springboot.app.services;
 import com.hostal.springboot.app.commons.GenericService;
 import com.hostal.springboot.app.model.Huesped;
 
-public interface HuespedService extends GenericService<Huesped, Long>{
+public interface HuespedService extends GenericService<Huesped, Integer>{
 
 }
