+++ src/main/java/com/hostal/springboot/app/services/ReservaHabitacionService.java
@@ -5,6 +5,6 @@ package com.hostal.springboot.app.services;
 import com.hostal.springboot.app.commons.GenericService;
 import com.hostal.springboot.app.model.ReservaHabitacion;
 
-public interface ReservaHabitacionService extends GenericService<ReservaHabitacion, Long> {
+public interface ReservaHabitacionService extends GenericService<ReservaHabitacion, Integer> {
 
 }
