+++ src/main/java/com/hostal/springboot/app/services/ReservaService.java
@@ -3,7 +3,7 @@ package com.hostal.springboot.app.services;
 import com.hostal.springboot.app.commons.GenericService;
 import com.hostal.springboot.app.model.Reserva;
 
-public interface ReservaService extends  GenericService<Reserva, Long>{
+public interface ReservaService extends  GenericService<Reserva, Integer>{
 	
 
 }
