+++ src/main/java/com/hostal/springboot/app/services/ServicioHabitacionService.java
@@ -3,6 +3,6 @@ package com.hostal.springboot.app.services;
 import com.hostal.springboot.app.commons.GenericService;
 import com.hostal.springboot.app.model.ServicioHabitacion;
 
-public interface ServicioHabitacionService extends GenericService<ServicioHabitacion, Long> {
+public interface ServicioHabitacionService extends GenericService<ServicioHabitacion, Integer> {
 
 }
