+++ src/main/java/com/hostal/springboot/app/services/ServicioService.java
@@ -3,6 +3,6 @@ package com.hostal.springboot.app.services;
 import com.hostal.springboot.app.commons.GenericService;
 import com.hostal.springboot.app.model.Servicio;
 
-public interface ServicioService extends GenericService<Servicio, Long> {
+public interface ServicioService extends GenericService<Servicio, Integer> {
 
 }
