+++ src/main/java/com/hostal/springboot/app/services/TipoPagoService.java
@@ -3,6 +3,6 @@ package com.hostal.springboot.app.services;
 import com.hostal.springboot.app.commons.GenericService;
 import com.hostal.springboot.app.model.TipoPago;
 
-public interface TipoPagoService extends GenericService<TipoPago, Long> {
+public interface TipoPagoService extends GenericService<TipoPago, Integer> {
 
 }
